med_reconciliation/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-med-reconciliation"
version = "0.1.0"
description = "Medication Reconciliation environment for OpenEnv — agent identifies dangerous drug discrepancies"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "med_reconciliation.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["med_reconciliation", "med_reconciliation.server", "med_reconciliation.graders"]
package-dir = { "med_reconciliation" = ".", "med_reconciliation.server" = "server", "med_reconciliation.graders" = "graders" }
